make source plain ascii
Non-ascii characters in the source could result in the following error
when importing the `neurio` module:

    Non-ASCII character '\xe2' in file neurio/__init__.py on line 447,
    but no encoding declared; see http://python.org/dev/peps/pep-0263/
    for details

This can be fixed either by replacing all the characters with plain
ascii equivalents or by adding a file header to specify the source file
encoding.

This commit implements the first option (replacing non-ascii characters
with ascii equivalents).